Rework: Use CfgFunctions
This removes the CBA dependency previously required for TB3 Gear, it also changes the function names used and how this is set up in a mission. Older versions will work but not in line with this.

Example unit also shows how you can have a unit start with magazines loaded by adding them as attachments.

Documentation changes to follow soon.